Port gtksourceview to GtkOSXApplication
This patch ports gtksourceview and gtksourceview3 to
GtkOSXApplication. They only use it to detect whether the program
is running in a bundle and if so, get the locale directory.
Gtksourceview3 also includes a patch for configure.ac that
fixes detecting whether to build for OS X or not.

See also bgo#664511, bgo#664679.